Add a TableHeader component to the Editor page.
This is a row that shows total amount and number of
transactions. It also has a check box to quickly select
all or deselect all transactions.

Change the style of the table to not have a top line
by default. Specific pages can add it if they
don't have a header.