Add 'spread' spending to the Daily tracker.
The idea is that for large purchases, you pay
it off over future days. For example, if you pay
rent monthly, you can add the tag spread:1m and
rather than being a single large cost, it will
be 1/30 of the cost for the next 30 days.

This makes it easier to pay for things like
vacations or other travel.