Ensure both files download successfully.
Previously, we were only checking the download
of the transactions.json file. It's possible that
either it would complete first or settings.json failed,
in which case, we would be in a broken state.

Now wait for both transactions.json and settings.json
to download successfully before rendering the page.